requirements.txt

fastapi
uvicorn
PyJWT
python-decouple
pydantic[email]
python-dotenv